auth/requirements.txt
Flask==3.1.0
PyJWT[crypto]==2.10.1